Visualisation exit functions

+++ examples/Keratinocyte/src/visualisation/visualisation.h
@@ -21,16 +21,16 @@ const unsigned int WINDOW_WIDTH = 512;
 const unsigned int WINDOW_HEIGHT = 512;
 
 //frustrum
-const double NEAR_CLIP = 0.1;
-const double FAR_CLIP = 1000;
+const float NEAR_CLIP = 0.1f;
+const float FAR_CLIP = 1000;
 
 //Circle model fidelity
 const int SPHERE_SLICES = 20;
 const int SPHERE_STACKS = 20;
-const double SPHERE_RADIUS = 10;
+const float SPHERE_RADIUS = 10;
 
 //Viewing Distance
-const double VIEW_DISTANCE = 500;
+const float VIEW_DISTANCE = 500;
 
 //light position
 GLfloat LIGHT_POSITION[] = {10.0f, 10.0f, 10.0f, 1.0f};
